src/requirements.txt
arabic-reshaper==2.1.3
certifi==2022.6.15
charset-normalizer==2.0.12
future==0.18.2
idna==3.3
Pillow==9.1.1
pyTelegramBotAPI==4.6.0
python-bidi==0.4.2
python-decouple==3.6
requests==2.28.0
six==1.16.0
urllib3==1.26.9
